+++ lib/constants/images.dart
@@ -1,4 +1,4 @@
 class HngImages{
   //sample
-static const  String homeImage = "assets/images/pfp3.jfif";
+static const  String homeImage = "assets/images/emoji.png";
 }
