+++ lib/screens/settings_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+class SettingsScreen extends StatefulWidget {
+  const SettingsScreen({super.key});
+
+  @override
+  State<SettingsScreen> createState() => _SettingsScreenState();
+}
+
+class _SettingsScreenState extends State<SettingsScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Text("Setting");
+  }
+}
+
